+++ CHANGELOG.md
@@ -24,6 +24,9 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
    - `MarkdownParserStateInterface`
    - `TableOfContentsGenerator`
    - `TableOfContentsGeneratorInterface`
+   - `TableOfContentsPlaceholder`
+   - `TableOfContentsPlaceholderParser`
+   - `TableOfContentsPlaceholderRenderer`
  - Added several new methods:
    - `FencedCode::setInfo()`
    - `HtmlRenderer::renderDocument()`
@@ -36,6 +39,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
    - `TableCell::getAlign()`
    - `TableCell::setAlign()`
  - Added the ability to properly clone `Node` objects
+ - Added the ability to render `TableOfContents` nodes anywhere in a document (given by a placeholder)
 
 ### Changed